% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapid-names.R
\name{.mapidnames}
\alias{.mapidnames}
\title{Internal function to map identifiers to names}
\usage{
.mapidnames(df, rebuild = FALSE)
}
\arguments{
\item{df}{Taxonomy Table converted to data frame format. There should be a column titled
\code{df_id} which represents the identifiers used in the original data set
(e.g. \code{taxa_names} from a \code{phyloseq})}

\item{rebuild}{(Logical). Indicate whether to re-download the NCBI database used to query}
}
\description{
Internal function to map identifiers to names
}
