% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapid-names.R
\name{.matching}
\alias{.matching}
\title{Perform proper matching using taxizedb and Jaccard distances}
\usage{
.matching(fullname, match, rank, ..., threshold = 0.1)
}
\arguments{
\item{fullname}{The full name separated by "_" for resolving ambiguous matching}

\item{match}{The column of the string to match and passed to taxizedb}
}
\description{
Perform proper matching using taxizedb and Jaccard distances
}
\keyword{internal}
