% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapid-generics.R
\name{mapid}
\alias{mapid}
\alias{mapid,phyloseq-method}
\alias{mapid,TreeSummarizedExperiment-method}
\title{Mapping taxonomic names to NCBI identifiers}
\usage{
mapid(obj, type, ..., rebuild_ncbi = FALSE, db = NULL)

\S4method{mapid}{phyloseq}(obj, type, ..., rebuild_ncbi = FALSE, db = NULL)

\S4method{mapid}{TreeSummarizedExperiment}(obj, type, ..., rebuild_ncbi = FALSE, db = NULL)
}
\arguments{
\item{obj}{A \code{phyloseq} type object or a \code{TreeSummarizedExperiment} type object}

\item{type}{Approach used to match names with identifiers. Can be either \code{db}, \code{path},
\code{name}.}

\item{db}{(String). If type is \code{db}, this argument specifies which db to use. Defaults to NULL.
Currently only supports \code{silva} and \code{metaphlan}}
}
\description{
This function takes a data container object (type \code{phyloseq}, \code{data.frame}\\\code{tibble},
\code{TreeSummarizedExperiment}) and returns NCBI identifiers.
}
\details{
The input is a \code{data.frame}, \code{matrix}, or typical microbiome container that has
a taxonomic table specified. The schematic of a taxonomic table is a table where the left-most
column indicates the highest level of the hierarchy (usually kingdom or superkingdom) and the right-most
column specifies the lowest level of the taxonomic hierarchy (usually species, genus or strains). Given
that format, the \code{type} argument determines how matching is done.
\itemize{
\item{\code{name}:}{ This matches using the name of the lowest rank in the taxonomic table to the NCBI database.}
\item{\code{path_*}:}{ Uses a full path matching algorithm as defined in Balvociute and Huson 2017.
(doi: 10.1186/s12864-017-3501-4)}
\item{\code{db}:}{ Use directly from a database that has a cross-walk between NCBI Ids and the full name path. Currently
only silva and metaphlan are supported. Future support might include RDP}.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{phyloseq}: \code{phyloseq} dispatch

\item \code{TreeSummarizedExperiment}: \code{TreeSummarizedExperiment} dispatch
}}

