% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_version}
\alias{check_version}
\title{Checking the version of each matching database}
\usage{
check_version(db = c("silva", "metaphlan"))
}
\arguments{
\item{db}{(Character). The name of the database.}
}
\value{
Returns version number as string.
}
\description{
Checking the version of each matching database
}
